fix bugs

+++ BombPlant.cpp
@@ -1,6 +1,6 @@
 #include "BombPlant.h"
 
-BombPlant::BombPlant(const size_t& cost, const size_t& hp) : Plant(cost, hp), damage_(hp) {}
+BombPlant::BombPlant(const int& cost, const int& hp) : Plant(cost, hp), damage_(hp) {}
 
 const int& BombPlant::myDamage() const
 {
