+++ BombPlant.h
@@ -8,7 +8,7 @@
 class BombPlant : public Plant
 {
     public:
-        BombPlant(const size_t& cost, const size_t& hp);
+        BombPlant(const int& cost, const int& hp);
         const int& myDamage() const;
         virtual const std::string& getTypeName() const override
         {
