+++ HornPlant.cpp
@@ -1,8 +1,8 @@
 #include "HornPlant.h"
 
-HornPlant::HornPlant(const size_t& cost, const size_t& hp, const size_t& damage) : Plant(cost, hp), damage_(damage) {}
+HornPlant::HornPlant(const int& cost, const int& hp, const int& damage) : Plant(cost, hp), damage_(damage) {}
 
-const size_t& HornPlant::getDamage() const
+const int& HornPlant::getDamage() const
 {
     return this->damage_;
 }
