+++ Plant.cpp
@@ -9,14 +9,14 @@ const std::array<std::string, 5> Plant::plantTypeNames =
     "Healflower"
 };
 
-// size_t Plant::max_hp_ = 0;
-// size_t Plant::count_of_plant_ = 0;
+// int Plant::max_hp_ = 0;
+// int Plant::count_of_plant_ = 0;
 
 Plant::Plant() : cost_(0), hp_(0) {}
 
 // manual: Plant plant(50, 75, PlantType::BOMB);
-// Plant::Plant(size_t cost, size_t hp, PlantType type) : cost_(cost), hp_(hp), type_(type) {}
-Plant::Plant(const size_t& cost, const size_t& hp) : cost_(cost), hp_(hp) {}
+// Plant::Plant(int cost, int hp, PlantType type) : cost_(cost), hp_(hp), type_(type) {}
+Plant::Plant(const int& cost, const int& hp) : cost_(cost), hp_(hp) {}
 
 Plant::~Plant()
 {
@@ -33,12 +33,12 @@ const int& Plant::getCost() const
     return this->cost_;
 }
 
-void Plant::beHealed(const size_t& healPoint)
+void Plant::beHealed(const int& healPoint)
 {
     this->hp_ += healPoint;
 }
 
-void Plant::beInjured(const size_t& injuryPoint)
+void Plant::beInjured(const int& injuryPoint)
 {
     this->hp_ -= injuryPoint;
 }
